Refactoring: Move error handler to BaseCommand
